Refactored Settings Service
