+++ App/app/src/main/java/com/wokabel/app/wokabel/SurvivalBoxApplication.java
@@ -1,0 +1,18 @@
+package com.wokabel.app.wokabel;
+
+import android.app.Application;
+
+import com.wokabel.app.wokabel.services.preferences.Settings;
+
+public class SurvivalBoxApplication extends Application {
+
+    public static Settings sharedPreferences;
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+        // Get Settings reference
+        sharedPreferences = Settings.getInstance(this);
+    }
+}
