+++ App/app/src/main/java/com/wokabel/app/wokabel/views/SubjectSelect.java
@@ -8,7 +8,8 @@ import android.view.MenuItem;
 import android.widget.TextView;
 
 import com.wokabel.app.wokabel.R;
-import com.wokabel.app.wokabel.models.Settings;
+import com.wokabel.app.wokabel.SurvivalBoxApplication;
+import com.wokabel.app.wokabel.services.preferences.Settings;
 
 public class SubjectSelect extends AppCompatActivity {
 
@@ -39,7 +40,7 @@ public class SubjectSelect extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_subject_select);
 
-        Settings settings = new Settings(getApplicationContext());
+        Settings settings = SurvivalBoxApplication.sharedPreferences;
         TextView username = findViewById(R.id.UsernameTextView);
         username.setText(getString(R.string.greet_user, settings.getString("username")));
